add validator test
